+++ api/src/api/v1/community_businesses/get.ts
@@ -3,7 +3,7 @@ import * as Boom from '@hapi/boom';
 import { pick } from 'ramda';
 import { CommunityBusinesses, CommunityBusiness } from '../../../models';
 import { getCommunityBusiness, isChildOrganisation } from '../prerequisites';
-import { GetCommunityBusinessRequest, GetCommunityBusinessesRequest } from '../types/requests';
+import { GetCommunityBusinessRequest, GetCommunityBusinessesRequest } from '../types/api';
 import { query, response } from './schema';
 import { is360GivingId } from '../prerequisites/get_community_business';
