+++ api/src/api/v1/community_businesses/register.ts
@@ -4,7 +4,7 @@ import { omit } from 'ramda';
 
 import { CommunityBusinesses, Users, CbAdmins } from '../../../models';
 import { isChildOrganisation } from '../prerequisites';
-import { RegisterCommunityBusinessesRequest } from '../types/requests';
+import { RegisterCommunityBusinessesRequest } from '../types/api';
 import { response, cbPayload } from './schema';
 import {
   userName,
