+++ api/src/api/v1/prerequisites/get_community_business.ts
@@ -10,7 +10,7 @@ import * as Hapi from '@hapi/hapi';
 import * as Boom from '@hapi/boom';
 import { isNil } from 'ramda';
 import { CommunityBusinesses } from '../../../models';
-import { GetCommunityBusinessRequest } from '../types/requests';
+import { GetCommunityBusinessRequest } from '../types/api';
 import { getCredentialsFromRequest } from '../auth';
