+++ api/src/api/v1/users/put.ts
@@ -16,7 +16,7 @@ import {
   ethnicity,
   response
 } from './schema';
-import { PutUserRequest } from '../types/requests';
+import { PutUserRequest } from '../types/api';
 import { isChildUser } from '../prerequisites';
 import { Credentials as StandardCredentials } from '../../../auth/strategies/standard';
