+++ api/src/api/v1/users/register/volunteers.ts
@@ -22,7 +22,7 @@ import {
   Users,
   CommunityBusinesses
 } from '../../../../models';
-import { VolunteerRegisterRequest } from '../../types/requests';
+import { VolunteerRegisterRequest } from '../../types/api';
 import { RoleEnum } from '../../../../models/types';
 import Roles from '../../../../models/role';
 import { Tokens } from '../../../../models/token';
